Extract login function to query module
The same login was function was used in two places.
